backend/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
pytesseract
opencv-python
boto3
pillow
openai
python-multipart
alembic>=1.10
pdf2image>=1.16.0
resend
jinja2
